fixed with dtos

+++ src/main/java/com/informatorio/finalproject/controller/AuthorController.java
@@ -1,5 +1,6 @@
 package com.informatorio.finalproject.controller;
 
+import com.informatorio.finalproject.converter.AuthorConverter;
 import com.informatorio.finalproject.dto.AuthorDTO;
 import com.informatorio.finalproject.entity.Author;
 import com.informatorio.finalproject.repository.AuthorRepository;
@@ -11,9 +12,11 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-
-import java.awt.print.Pageable;
-import java.util.Collection;
+import javax.validation.Valid;
+import java.util.HashMap;
+import java.util.Map;
+import java.util.Optional;
+import java.util.stream.Collectors;
 
 @RestController
 public class AuthorController {
@@ -21,35 +24,64 @@ public class AuthorController {
     private AuthorRepository authorRepository;
     private final AuthorService authorService;
 
+    private final AuthorConverter authorConverter;
+
     @Autowired
-    public AuthorController(AuthorRepository authorRepository, AuthorService authorService){
+    public AuthorController(AuthorRepository authorRepository, AuthorService authorService, AuthorConverter authorConverter){
         this.authorRepository = authorRepository;
         this.authorService = authorService;
+        this.authorConverter = authorConverter;
     }
 
     @PostMapping("/authors")
-    public ResponseEntity<?> saveAuthor(@RequestBody AuthorDTO authorDTO){
-        Author author = new Author(null, authorDTO.getFirstName(), authorDTO.getLastName(), authorDTO.getFullName(), authorDTO.getCreatedAt());
-        return new ResponseEntity<>(authorRepository.save(author),HttpStatus.CREATED);
+    public ResponseEntity<?> createAuthor(@Valid @RequestBody AuthorDTO authorDTO){
+        Author author = authorConverter.toEntity(authorDTO);
+        author = authorRepository.save(author);
+        return new ResponseEntity<>(authorConverter.toDto(author), HttpStatus.CREATED);
     }
 
     @GetMapping("/authors")
     public ResponseEntity<?>getAuthors(@RequestParam int page){
         PageRequest pageable = PageRequest.of(page, 3);
         Page<Author> pageResult = authorRepository.findAll(pageable);
-        return new ResponseEntity<>(pageResult, HttpStatus.OK);
+
+        Map<String, Object> customPage = new HashMap<>();
+        customPage.put("content", pageResult.getContent().stream()
+                .map(author -> authorConverter.toDto(author))
+                .collect(Collectors.toList()));
+        customPage.put("page", pageResult.getNumber());
+        customPage.put("size", pageResult.getSize());
+        customPage.put("totalElements", pageResult.getTotalElements());
+
+        return new ResponseEntity<>(customPage, HttpStatus.OK);
 
     }
 
     @GetMapping("/authors/{id}")
-    public ResponseEntity<?>getAuthors(@PathVariable Long id){
-        return ResponseEntity.ok(authorRepository.findById(id));
+    public ResponseEntity<?>findById(@PathVariable Long id){
+        Optional<Author> author = authorRepository.findById(id);
+        if (author.isPresent()){
+            return new ResponseEntity<>(authorConverter.toDto(author.get()), HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
     }
 
     @DeleteMapping(value = "/authors/{id}")
-    public void deleteAuthor(@PathVariable Long id){
-        authorRepository.deleteById(id);
+    public ResponseEntity<?> deleteAuthor(@PathVariable Long id){
+        Optional<Author> author = authorRepository.findById(id);
+        if (author.isPresent()){
+            authorRepository.deleteById(id);
+            return new ResponseEntity<>(HttpStatus.ACCEPTED);
+        }else {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
     }
+
+   /*@PutMapping("/authors")
+    public Author updateAuthor(@RequestBody AuthorDTO authorDTO){
+        return authorService.updateAuthor(authorDTO);
+    }*/
